Add canvas setup, shape creation, and event handling hooks: implement canvas initialization, shape drawing, and user interaction features for enhanced canvas functionality
